+++ test/Spec.hs
@@ -1,3 +1,3 @@
--- {-# OPTIONS_GHC -F -pgmF doctest-discover #-}
-main :: IO ()
-main = putStrLn "Only doctest suite is implemented, no other tests"
+{-# OPTIONS_GHC -F -pgmF hspec-discover #-}
+-- main :: IO ()
+-- main = putStrLn "Only doctest suite is implemented, no other tests"